refactor: extract JsBaseTableProxy and add JsTreeTableProxy
Split shared table proxy logic into a generic JsBaseTableProxy base class and add a dedicated JsTreeTableProxy for TreeTable (including rollup) so TreeTables no longer reuse JsTableProxy.
